Move tag group and `tags.py` into a single module
