+++ src/extensions/util.py
@@ -138,287 +138,6 @@ class Util(commands.Cog):
         ).set_image(url=user.display_avatar.url))
 
 
-    # tag
-    # The custom app_commands.Group class exists so I can 
-    # apply this decorator
-    @app_commands.guild_only()
-    class TagGroup(app_commands.Group):
-        pass
-    tag_group = TagGroup(name='tag', description='Añade o usa tags tuyos o de otros usuarios')
-
-
-    # tag show
-    @tag_group.command(name='show')
-    @app_commands.rename(tag_name='tag')
-    async def tag_show(self, interaction: discord.Interaction, tag_name: app_commands.Range[str, 1, 32]):
-        """Muestra el contenido de un tag
-
-        tag_name
-            Nombre de un tag
-        """
-        await tags.tag_check(interaction)
-        tag = tags.get_tag(interaction, tag_name)
-        if not (not interaction.channel.is_nsfw() and bool(tag.nsfw)):
-            await interaction.response.send_message(await commands.clean_content().convert(interaction, tag.content))
-        else:
-            await interaction.response.send_message(core.Warning.error('Este tag solo puede mostrarse en canales NSFW'), ephemeral=True)
-
-
-    # tag toggle
-    @app_commands.checks.cooldown(1, 10, key=lambda i: i.guild_id)
-    @tag_group.command(name='toggle')
-    async def tag_toggle(self, interaction: discord.Interaction):
-        """Activa los tags en el servidor"""
-        if interaction.channel.permissions_for(interaction.user).manage_guild:
-            db.cursor.execute("SELECT guild FROM tagsenabled WHERE guild=?", (interaction.guild_id,))
-            check: tuple[int] | None = db.cursor.fetchone()
-            if check is None:
-                confirmation = core.Confirm(interaction, interaction.user)
-                await interaction.response.send_message(core.Warning.question('Los tags en este servidor están desactivados. ¿Quieres activarlos?'), view=confirmation, ephemeral=True)
-                await confirmation.wait()
-
-                if confirmation.value is None:
-                    return
-
-                confirmation.clear_items()
-
-                if confirmation.value:
-                    db.cursor.execute("INSERT INTO tagsenabled VALUES(?)", (interaction.guild_id,))
-                    await confirmation.last_interaction.response.edit_message(content=core.Warning.success('Se activaron los tags en este servidor'), view=confirmation)
-
-                else:
-                    await confirmation.last_interaction.response.edit_message(content=core.Warning.cancel('No se activaran los tags en este servidor'), view=confirmation)
-
-            else:
-                confirmation = core.Confirm(interaction, interaction.user)
-                await interaction.response.send_message(core.Warning.question('Los tags en este servidor están activados. ¿Quieres desactivarlos?'), view=confirmation, ephemeral=True)
-                await confirmation.wait()
-
-                if confirmation.value is None:
-                    return
-
-                confirmation.clear_items()
-
-                if confirmation.value:
-                    db.cursor.execute("DELETE FROM tagsenabled WHERE guild=?", (interaction.guild_id,))
-                    await confirmation.last_interaction.response.edit_message(content=core.Warning.success('Se desactivaron los tags en este servidor'), view=confirmation)
-
-                else:
-                    await confirmation.last_interaction.response.edit_message(content=core.Warning.cancel('No se desactivarán los tags en este servidor'), view=confirmation)
-            
-            db.conn.commit()
-
-        else:
-            await interaction.response.send_message(core.Warning.error('Necesitas permiso de gestionar servidor para activar o desactivar los tags'), ephemeral=True)
-
-
-    # tag add
-    @app_commands.checks.cooldown(1, 10)
-    @tag_group.command(name='add')
-    @app_commands.rename(tag_name='nombre', tag_content='contenido')
-    async def tag_add(
-            self,
-            interaction: discord.Interaction, 
-            tag_name: app_commands.Range[str, 1, 32], 
-            tag_content: str,
-            nsfw: bool = False
-        ):
-        """Crea un tag
-
-        tag_name
-            Nombre del tag que quieres crear
-        tag_content
-            Contenido del tag que quieres crear
-        nsfw
-            Determina si el tag puede mostrarse únicamente en canales NSFW
-        """
-        await tags.tag_check(interaction)
-        tags.check_tag_name(interaction, tag_name)
-        if interaction.channel.nsfw:
-            nsfw = True
-        tags.add_tag(interaction, tag_name, tag_content, nsfw)
-        await interaction.response.send_message(core.Warning.success(f'Se agregó el tag **{await commands.clean_content().convert(interaction, tag_name)}**'))
-
-
-    # tag gift
-    @app_commands.checks.cooldown(1, 10)
-    @tag_group.command(name='gift')
-    @app_commands.rename(tag_name='tag', user='usuario')
-    async def tag_gift(
-            self,
-            interaction: discord.Interaction, 
-            tag_name: app_commands.Range[str, 1, 32],
-            user: discord.Member
-        ):
-        """Regala un tag a otro usuario"""
-        await tags.tag_check(interaction)
-        if user == interaction.user:
-            await interaction.response.send_message(core.Warning.error('No puedes regalarte un tag a ti mismo'), ephemeral=True)
-        elif user.bot:
-            await interaction.response.send_message(core.Warning.error('No puedes regalarle un tag a un bot'), ephemeral=True)
-        else:
-            tag: tags.Tag = tags.get_tag(interaction, tag_name)
-            if tag.user.id != interaction.user.id:
-                await interaction.response.send_message(core.Warning.error('No puedes regalar el tag de otra persona'), ephemeral=True)
-                return
-            gift_permission = core.Confirm(interaction, user)
-            await interaction.response.send_message(core.Warning.question(f'{user.mention} ¿Quieres aceptar el tag **{await commands.clean_content().convert(interaction, tag.name)}** por parte de {interaction.user.name}?'), view=gift_permission)
-            await gift_permission.wait()
-
-            if gift_permission.value is None:
-                return
-            
-            gift_permission = gift_permission.clear_items()
-
-            if gift_permission.value:
-                tag.gift(user)
-                await gift_permission.last_interaction.response.edit_message(content=core.Warning.success(f'El tag **{await commands.clean_content().convert(interaction, tag.name)}** fue regalado a {await commands.clean_content().convert(interaction, tag.user.name)} por parte de {await commands.clean_content().convert(interaction, interaction.user.name)}'), view=gift_permission)
-
-            elif not gift_permission.value:
-                await gift_permission.last_interaction.response.edit_message(content=core.Warning.cancel(f'{interaction.user.mention} El regalo fue rechazado'), view=gift_permission)
-
-
-    # tag rename
-    @app_commands.checks.cooldown(1, 5)
-    @tag_group.command(name='rename')
-    @app_commands.rename(old_name='tag', new_name='nuevo')
-    async def tag_rename(
-            self,
-            interaction: discord.Interaction,
-            old_name: app_commands.Range[str, 1, 32],
-            new_name: app_commands.Range[str, 1, 32]
-        ):
-        """Cambia el nombre de uno de tus tags"""
-        await tags.tag_check(interaction)
-        tags.check_tag_name(interaction, new_name)
-        tag = tags.get_tag(interaction, old_name)
-        if tag.user.id == interaction.user.id:
-            if tag.name == new_name:
-                await interaction.response.send_message(core.Warning.error('No puedes ponerle el mismo nombre a un tag'), ephemeral=True)
-            else:
-                tag.rename(new_name)
-                await interaction.response.send_message(core.Warning.success(f'El nombre del tag **{await commands.clean_content().convert(interaction, old_name)}** se ha cambiado a **{await commands.clean_content().convert(interaction, new_name)}**'))
-
-        else:
-            await interaction.response.send_message(core.Warning.error('No puedes renombarar el tag de otro usuario'), ephemeral=True)
-
-
-    # tag edit
-    @app_commands.checks.cooldown(1, 10)
-    @tag_group.command(name='edit')
-    @app_commands.rename(tag_name='tag', tag_content='contenido')
-    async def tag_edit(
-            self,
-            interaction: discord.Interaction,
-            tag_name: app_commands.Range[str, 1, 32],
-            tag_content: str,
-            nsfw: bool = False
-        ):
-        """Edita el contenido de uno de tus tags
-
-        tag_name
-            Nombre del tag que quieres editar
-        tag_content
-            Nuevo contenido del tag
-        nsfw
-            Determina si el tag puede mostrarse únicamente en canales NSFW
-        """
-        await tags.tag_check(interaction)
-        tag = tags.get_tag(interaction, tag_name)
-        if interaction.user.id == tag.user.id:
-            if interaction.channel.nsfw:
-                nsfw = True
-            tag.edit(tag_content, nsfw)
-            await interaction.response.send_message(core.Warning.success(f'Se editó el tag **{await commands.clean_content().convert(interaction, tag_name)}**'))
-        
-        else:
-            await interaction.response.send_message(core.Warning.error('No puedes editar tags de otros usuarios'), ephemeral=True)
-
-
-    # tag delete
-    @app_commands.checks.cooldown(1, 5)
-    @tag_group.command(name='delete')
-    @app_commands.rename(tag_name='tag')
-    async def tag_delete(self, interaction: discord.Interaction, tag_name: app_commands.Range[str, 1, 32]):
-        """Elimina uno de tus tags"""
-        await tags.tag_check(interaction)
-        tag = tags.get_tag(interaction, tag_name)
-        if interaction.user.id == tag.user.id:
-            confirmation = core.Confirm(interaction, interaction.user)
-            await interaction.response.send_message(core.Warning.question(f'¿Quieres eliminar el tag **{await commands.clean_content().convert(interaction, tag_name)}**?'), view=confirmation)
-            await confirmation.wait()
-            
-            if confirmation.value is None:
-                return
-            
-            confirmation = confirmation.clear_items()
-
-            if confirmation.value:
-                tag.delete()
-                await confirmation.last_interaction.response.edit_message(content=core.Warning.success(f'El tag **{await commands.clean_content().convert(interaction, tag_name)}** ha sido eliminado'), view=confirmation)
-            
-            elif not confirmation.value:
-                await confirmation.last_interaction.response.edit_message(content=core.Warning.cancel(f'El tag no será eliminado'), view=confirmation)
-
-        else: 
-            await interaction.response.send_message(core.Warning.error('No puedes eliminar tags de otros usuarios'), ephemeral=True)
-
-
-    # tag forcedelete
-    @tag_group.command(name='forcedelete')
-    @core.owner_only()
-    async def forcedelete(
-            self,
-            interaction: discord.Interaction,
-            tag_name: app_commands.Range[str, 1, 32],
-            guild_id: str | None,
-            silent: bool = False
-        ):
-        """Reservado"""
-        guild = interaction.guild if guild_id is None else self.bot.get_guild(int(guild_id))
-        tag = tags.get_tag(interaction, tag_name, guild)
-        tag.delete()
-        await interaction.response.send_message(core.Warning.success(f'El tag **{await commands.clean_content().convert(interaction, tag_name)}** ha sido eliminado'), ephemeral=silent)
-
-
-    # tag owner
-    @app_commands.checks.cooldown(1, 3)
-    @tag_group.command(name='owner')
-    @app_commands.rename(tag_name='tag')
-    async def tag_owner(self, interaction: discord.Interaction, tag_name: app_commands.Range[str, 1, 32]):
-        """Muestra el propietario de un tag"""
-        await tags.tag_check(interaction)
-        tag = tags.get_tag(interaction, tag_name)
-        await interaction.response.send_message(core.Warning.info(f'El dueño del tag **{await commands.clean_content().convert(interaction, tag.name)}** es `{await commands.clean_content().convert(interaction, str(tag.user))}`'))
-
-
-    # tag list
-    @app_commands.checks.cooldown(1, 10)
-    @tag_group.command(name='list')
-    @app_commands.rename(user='usuario')
-    async def tag_list(self, interaction: discord.Interaction, user: discord.Member | None):
-        """Muestra una lista de tus tags o de los tags de otro usuario"""
-        await tags.tag_check(interaction)
-        if user is None:
-            user = interaction.user
-        tag_list = list(map(lambda tag: f'"{tag}"', tags.get_member_tags(interaction, user, raises=True)))
-        pages = pagination.Page.from_list(interaction, f'Tags de {user.name}', tag_list)
-        paginator = pagination.Paginator.optional(interaction, pages=pages, entries=len(tag_list))
-        await interaction.response.send_message(embed=pages[0].embed, view=paginator)
-
-
-    # tag serverlist
-    @app_commands.checks.cooldown(1, 20)
-    @tag_group.command(name='serverlist')
-    async def tag_serverlist(self, interaction: discord.Interaction):
-        """Muestra los tags de todo el servidor"""
-        await tags.tag_check(interaction)
-        tag_list = list(map(lambda tag: f'{tag.user.name}: "{tag}"', tags.get_guild_tags(interaction, raises=True)))
-        pages = pagination.Page.from_list(interaction, f'Tags de {interaction.guild}', tag_list)
-        paginator = pagination.Paginator.optional(interaction, pages=pages, entries=len(tag_list))
-        await interaction.response.send_message(embed=pages[0].embed, view=paginator)
-
-
     # someone
     @app_commands.checks.cooldown(3, 10)
     @app_commands.command()
